feat(handshake): refactor handshake implementation into client and server modules
