+++ libzappy_net/src/zappy_net_internal.h
@@ -11,6 +11,9 @@
     #include "../include/zappy_net.h"
     #include "../include/zappy_net_ringbuf.h"
     #include <netinet/in.h>
+    #include <string.h>
+    #include <stdio.h>
+    #include <stdlib.h>
 
 /**
  * @brief Internal socket structure definition
@@ -24,5 +27,11 @@ struct zn_socket {
     zn_ringbuf_t write_buffer; /* Buffer for writing to socket */
     int buffer_initialized;    /* Flag to track if buffers are initialized */
 };
+
+/**
+ * @brief Internal handshake utility functions
+ */
+int zn_send_message(zn_socket_t sock, const char *message);
+char *zn_receive_message(zn_socket_t sock);
 
 #endif /* !ZAPPY_NET_INTERNAL_H_ */
